fix(rev-245.2): audit G6 — bzip2 carry typing, Latin-1 desc bytes, LruCache doc lineage
- bzip2-01/02: total_in/out_lo32+hi32 typed int32 so the 32-bit carry
  idioms (lo32<prev → hi32++, lo32==0 → hi32++) detect wraparound as in
  Java (BZip2.java:160-163,432-435). Write-only telemetry, faithful typing.
- objtype-02: ToCertificate desc encodes one byte per char like Java
  String.getBytes() (ObjType.java:398); []byte(string) emitted UTF-8
  multi-byte sequences that the Latin1→UTF8 consumer garbled for
  non-ASCII names.
- datastruct-07: no code change — Java's duplicate-key put behavior is
  unreproducible on a Go map without bucket chains; the documented
  Get-then-Put-on-miss caller constraint stands, 2026-06-04 audit
  re-confirmation added to the Put comment.
- loctype-02: no change — the codes-35-38 container overflow is faithful
  upstream Java behavior (bug-for-bug parity).

+++ pkg/jagex2/io/bzip2/bzip2.go
@@ -160,7 +160,7 @@ label67:
 	}
 
 	var13 := s.TotalOutLo32
-	s.TotalOutLo32 += availOutInit - csAvailOut
+	s.TotalOutLo32 += int32(availOutInit - csAvailOut)
 	if s.TotalOutLo32 < var13 {
 		s.TotalOutHi32++
 	}
